+++ apps/web/public/sw/activate.js
@@ -1,0 +1,18 @@
+import { CACHE_NAME } from './cache.js';
+
+self.addEventListener('activate', event => {
+  const cacheWhitelist = [CACHE_NAME];
+
+  event.waitUntil(
+    caches.keys().then(cacheNames => {
+      return Promise.all(
+        cacheNames.map(cacheName => {
+          if (!cacheWhitelist.includes(cacheName)) {
+            return caches.delete(cacheName); // Delete old caches
+          }
+        }),
+      );
+    }),
+  );
+  console.log('Service Worker activated.');
+});